+++ xbmc/addons/gui/CMakeLists.txt
@@ -1,8 +1,10 @@
 set(SOURCES GUIDialogAddonInfo.cpp
+            GUIDialogAddonSettings.cpp
             GUIViewStateAddonBrowser.cpp
             GUIWindowAddonBrowser.cpp)
 
 set(HEADERS GUIDialogAddonInfo.h
+            GUIDialogAddonSettings.h
             GUIViewStateAddonBrowser.h
             GUIWindowAddonBrowser.h)
