+++ xbmc/addons/gui/GUIDialogAddonSettings.cpp
