+++ xbmc/addons/gui/GUIDialogAddonSettings.h
