+++ xbmc/addons/settings/CMakeLists.txt
@@ -1,9 +1,7 @@
 set(SOURCES AddonSettings.cpp
-            GUIDialogAddonSettings.cpp
             SettingUrlEncodedString.cpp)
 
 set(HEADERS AddonSettings.h
-            GUIDialogAddonSettings.h
             SettingUrlEncodedString.h)
 
 core_add_library(addons_settings)
