+++ xbmc/games/dialogs/osd/DialogGameAdvancedSettings.cpp
@@ -10,7 +10,7 @@
 
 #include "ServiceBroker.h"
 #include "addons/AddonManager.h"
-#include "addons/settings/GUIDialogAddonSettings.h"
+#include "addons/gui/GUIDialogAddonSettings.h"
 #include "cores/RetroPlayer/guibridge/GUIGameRenderManager.h"
 #include "cores/RetroPlayer/guibridge/GUIGameSettingsHandle.h"
 #include "guilib/GUIMessage.h"
