+++ xbmc/music/dialogs/GUIDialogInfoProviderSettings.cpp
@@ -11,8 +11,8 @@
 #include "ServiceBroker.h"
 #include "Util.h"
 #include "addons/AddonSystemSettings.h"
+#include "addons/gui/GUIDialogAddonSettings.h"
 #include "addons/gui/GUIWindowAddonBrowser.h"
-#include "addons/settings/GUIDialogAddonSettings.h"
 #include "dialogs/GUIDialogFileBrowser.h"
 #include "dialogs/GUIDialogKaiToast.h"
 #include "filesystem/AddonsDirectory.h"
