+++ xbmc/settings/dialogs/GUIDialogContentSettings.cpp
@@ -10,8 +10,8 @@
 
 #include "ServiceBroker.h"
 #include "addons/AddonSystemSettings.h"
+#include "addons/gui/GUIDialogAddonSettings.h"
 #include "addons/gui/GUIWindowAddonBrowser.h"
-#include "addons/settings/GUIDialogAddonSettings.h"
 #include "dialogs/GUIDialogKaiToast.h"
 #include "dialogs/GUIDialogSelect.h"
 #include "filesystem/AddonsDirectory.h"
